Fix input params types. Optimize BeautifulQuadruples

+++ src/DataStructuresAlgorithms.h
@@ -342,7 +342,7 @@ vector<int> freqQuery(vector<vector<int>> &);
 string encryption(string &);
 vector<size_t> ClimbLeaderBoard(vector<long> &, vector<long> &);
 long CalculateMedian(vector<long> &);
-string TimeInWords(int, int);
+string TimeInWords(size_t, size_t);
 size_t BeautifulQuadruples(long, long, long, long);
 long kruskals(int, vector<long> &, vector<long> &, vector<long> &);
 long getLowestPathCost(size_t g_nodes, vector<long> &, vector<long> &, vector<long> &);
